+++ CesiumGltf/include/CesiumGltf/Model.h
@@ -142,9 +142,9 @@ struct CESIUMGLTF_API Model : public ModelSpec {
    * @brief Adds an extension to the {@link ModelSpec::extensionsUsed}
    * property, if it is not already present.
    *
-   * @param extensionUsed The name of the used extension.
+   * @param extensionName The name of the used extension.
    */
-  void addExtensionUsed(const std::string& extensionUsed);
+  void addExtensionUsed(const std::string& extensionName);
 
   /**
    * @brief Adds an extension to the {@link ModelSpec::extensionsRequired}
@@ -153,8 +153,27 @@ struct CESIUMGLTF_API Model : public ModelSpec {
    * Calling this function also adds the extension to `extensionsUsed`, if it's
    * not already present.
    *
-   * @param extensionRequired The name of the required extension.
+   * @param extensionName The name of the required extension.
    */
-  void addExtensionRequired(const std::string& extensionRequired);
+  void addExtensionRequired(const std::string& extensionName);
+
+  /**
+   * @brief Determines whether a given extension name is listed in the model's
+   * {@link ModelSpec::extensionsUsed} property.
+   *
+   * @param extensionName The extension name to check.
+   * @returns True if the extension is found in `extensionsUsed`; otherwise, false.
+   */
+  bool isExtensionUsed(const std::string& extensionName) const noexcept;
+
+  /**
+   * @brief Determines whether a given extension name is listed in the model's
+   * {@link ModelSpec::extensionsRequired} property.
+   *
+   * @param extensionName The extension name to check.
+   * @returns True if the extension is found in `extensionsRequired`; otherwise, false.
+   */
+  bool isExtensionRequired(const std::string& extensionName) const noexcept;
 };
+
 } // namespace CesiumGltf