+++ CHANGES.md
@@ -11,7 +11,7 @@
 - Added `Uri::getPath` and `Uri::setPath`.
 - Added `TileTransform::setTransform`.
 - Added a new `CesiumLegacyTerrain` library and namespace, containing classes for working with terrain in the `quantized-mesh-1.0` format and its `layer.json` file.
-- Added `addExtensionUsed` and `addExtensionRequired` methods to `CesiumGltf::Model`.
+- Added `addExtensionUsed`, `addExtensionRequired`, `isExtensionUsed`, and `isExtensionRequired` methods to `CesiumGltf::Model`.
 
 ##### Fixes :wrench: